refactor: abstract single table header component
